feat: OSX Support
